Fix SNR calculate error, total_snrlevels, requirements.txt

+++ noisyspeech_synthesizer.py
@@ -12,7 +12,7 @@ from audiolib import audioread, audiowrite, snr_mixer
 def main(cfg):
     snr_lower = float(cfg["snr_lower"])
     snr_upper = float(cfg["snr_upper"])
-    total_snrlevels = float(cfg["total_snrlevels"])
+    total_snrlevels = int(cfg["total_snrlevels"])
     
     clean_dir = os.path.join(os.path.dirname(__file__), 'clean_train')
     if cfg["speech_dir"]!='None':
